pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "xorq-datafusion"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "pyarrow>=19.0.1",
]

[dependency-groups]
dev = [
    "codespell[hard-encoding-detection,toml]==2.4.1",
    "pre-commit==4.2.0",
    "pytest==8.3.5",
    "maturin>=1.7.4",
    "pip>=24.3.1",
    "ruff>=0.11.2",
    "pytest-cov>=6.0.0",
    "coverage[toml]==7.8.0",
    "ipython>=8.34.0",
]
test = [
    "ibis-framework[duckdb]>=10.4.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
]

[tool.maturin]
module-name = "xorq_datafusion._internal"
python-source = "python"
features = ["pyo3/extension-module"]
